requirements-travis.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.1
django-allauth==0.42.0
django-bootstrap4==2.2.0
django-heroku==0.3.1
django-money==1.1
django-tables2==2.3.3
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
oauthlib==3.1.0
Pillow==8.0.1
psycopg2==2.8.6
py-moneyed==0.8.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
zipp==3.1.0
